requirements.txt
google-cloud-datastore==1.9.0
google-cloud-storage==1.19.1
python-dotenv==0.10.3
python-json-logger==0.1.11
